@@ -5,7 +5,7 @@
   "scripts": {
     "start": "next-remote-watch ./data",
     "dev": "next dev",
-    "build": "next build && node ./scripts/generate-sitemap",
+    "build": "next build",
     "serve": "next start",
     "analyze": "cross-env ANALYZE=true next build",
     "lint": "next lint --fix --dir pages --dir components --dir lib --dir layouts --dir scripts",
@@ -18,30 +18,32 @@
     "esbuild": "^0.12.15",
     "gray-matter": "^4.0.2",
     "image-size": "1.0.0",
-    "mdx-bundler": "^5.1.2",
-    "next": "11.0.1",
+    "mdx-bundler": "^6.0.1",
+    "next": "11.1.0",
     "next-themes": "^0.0.14",
     "postcss": "^8.3.5",
     "preact": "^10.5.13",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "reading-time": "1.3.0",
-    "rehype-katex": "^5.0.0",
-    "rehype-prism-plus": "0.0.4",
-    "remark-autolink-headings": "6.0.1",
-    "remark-footnotes": "^3.0.0",
-    "remark-gfm": "^1.0.0",
-    "remark-math": "^4.0.0",
-    "remark-slug": "6.0.0",
-    "tailwindcss": "^2.2.2"
+    "rehype-katex": "^6.0.0",
+    "rehype-prism-plus": "^0.0.5",
+    "remark-autolink-headings": "^7.0.0",
+    "remark-footnotes": "^4.0.0",
+    "remark-gfm": "^2.0.0",
+    "remark-math": "^5.0.0",
+    "remark-slug": "^7.0.0",
+    "sharp": "^0.28.3",
+    "tailwindcss": "^2.2.2",
+    "unist-util-visit": "^4.0.0"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "11.0.1",
+    "@next/bundle-analyzer": "11.1.0",
     "@svgr/webpack": "^5.5.0",
     "cross-env": "^7.0.3",
     "dedent": "^0.7.0",
     "eslint": "^7.29.0",
-    "eslint-config-next": "11.0.1",
+    "eslint-config-next": "11.1.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-prettier": "^3.3.1",
@@ -51,13 +53,7 @@
     "inquirer": "^8.1.1",
     "lint-staged": "^11.0.0",
     "next-remote-watch": "^1.0.0",
-    "prettier": "2.2.1",
-    "rehype": "11.0.0",
-    "remark-frontmatter": "3.0.0",
-    "remark-parse": "9.0.0",
-    "remark-stringify": "9.0.1",
-    "unified": "9.2.1",
-    "unist-util-visit": "2.0.3"
+    "prettier": "2.2.1"
   },
   "lint-staged": {
     "*.+(js|jsx|ts|tsx)": [
